recipe_nlp/requirements-dev.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
extruct==0.18.0
fastapi==0.115.6
fastapi-cli==0.0.7
flexcache==0.3
flexparser==0.4
h11==0.14.0
html5lib==1.1
html_text==0.6.2
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
inflect==7.5.0
ingredient_parser_nlp==1.3.2
iniconfig==2.0.0
isodate==0.7.2
Jinja2==3.1.5
joblib==1.4.2
jstyleson==0.0.2
lxml==5.3.0
lxml_html_clean==0.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mf2py==2.0.1
more-itertools==10.5.0
nltk==3.9.1
packaging==24.2
Pint==0.24.4
platformdirs==4.3.6
pluggy==1.5.0
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
pyRdfa3==3.6.4
pytest==8.3.4
python-crfsuite==0.9.11
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rdflib==7.1.1
recipe_scrapers==15.3.3
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.12.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
tqdm==4.67.1
typeguard==4.4.1
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
w3lib==2.2.1
watchfiles==1.0.3
webencodings==0.5.1
websockets==14.1
